refactor: wasm module handlers
