+++ src/mappings/wasm/index.ts
@@ -1,0 +1,3 @@
+export * from "./bridge";
+export * from "./contracts";
+export * from "./cw20";
